@W-14131330@ Refactor auth storage (#1462)
* Split storage into multiple files.

* Move "available" detection from base class to subclasses.

* Add utils to detect local storage and cookies.

* Clean up store init logic.

* Use `return await` for better call stacks.

* Add check for `navigator.cookieEnabled`.

* Add explainer for dispatching a storage event.

* Make it easier to start retail react app from monorepo root.

* Use safer local storage detection.

* Use local storage hook in retail react app.

* Clean up logic.

* Add comment and remove import.

* Revert aggressive storage detection until we can handle it better.

* Revert refinements changes.

* Finish sentence...

* Extract shared code to function.
