+++ packages/commerce-sdk-react/src/auth/storage/index.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) 2023, Salesforce, Inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: BSD-3-Clause
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+export type StorageType = 'cookie' | 'local' | 'memory'
+export * from './base'
+export * from './cookie'
+export * from './local'
+export * from './memory'
